+++ src/common/config/typeorm.config.ts
@@ -1,5 +1,12 @@
-import * as dotenv  from 'dotenv';
-dotenv.config()
+import * as dotenv from 'dotenv';
+dotenv.config();
+
+import { Admin } from '../../modules/admin/admin.entity';
+import { City } from '../../modules/city/city.entity';
+import { Client } from '../../modules/client/client.entity';
+import { ClockSize } from '../../modules/clockSize/clockSize.entity';
+import { Employee } from '../../modules/employee/employee.entity';
+import { Reservation } from '../../modules/reservation/reservation.entity';
 
 export = {
   host: process.env.HOST,
@@ -9,11 +16,11 @@ export = {
   password: process.env.PASSWORD,
   database: process.env.DATABASE,
   synchronize: false,
-  entities: ['src/entity/**/*.ts'],
+  entities: [Admin, City, Client, ClockSize, Employee, Reservation],
   migrations: ['src/migration/**/*.ts'],
   cli: {
     migrationsDir: 'src/migration',
     entitiesDir: 'src/entity',
     subscribersDir: 'src/subscriber',
   },
-}
+};
